fix: remove event package, shuffling things around.
