+++ src/main/java/com/wanderersoftherift/wotr/client/render/RenderEvents.java
@@ -1,9 +1,8 @@
-package com.wanderersoftherift.wotr.events;
+package com.wanderersoftherift.wotr.client.render;
 
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.mojang.blaze3d.vertex.VertexConsumer;
 import com.wanderersoftherift.wotr.WanderersOfTheRift;
-import com.wanderersoftherift.wotr.client.render.ModelWireframeExtractor;
 import it.unimi.dsi.fastutil.objects.Reference2ObjectOpenHashMap;
 import net.minecraft.client.Camera;
 import net.minecraft.client.Minecraft;
