--- src/main/java/com/wanderersoftherift/wotr/events/client/ModClientEvents.java
@@ -1,37 +1,0 @@
-package com.wanderersoftherift.wotr.events.client;
-
-import com.mojang.blaze3d.platform.InputConstants;
-import com.wanderersoftherift.wotr.WanderersOfTheRift;
-import net.minecraft.client.KeyMapping;
-import net.minecraft.client.Minecraft;
-import net.neoforged.api.distmarker.Dist;
-import net.neoforged.bus.api.SubscribeEvent;
-import net.neoforged.fml.common.EventBusSubscriber;
-import net.neoforged.fml.event.lifecycle.FMLClientSetupEvent;
-import net.neoforged.neoforge.client.event.RegisterKeyMappingsEvent;
-import org.lwjgl.glfw.GLFW;
-
-@EventBusSubscriber(modid = WanderersOfTheRift.MODID, bus = EventBusSubscriber.Bus.MOD, value = Dist.CLIENT)
-public class ModClientEvents {
-
-    public static final KeyMapping JIGSAW_NAME_TOGGLE_KEY = new KeyMapping(
-            "key." + WanderersOfTheRift.id("jigsaw_name_toggle"),
-            InputConstants.Type.KEYSYM,
-            GLFW.GLFW_KEY_UNKNOWN,
-            "key.categories.misc"
-    );
-
-
-    @SubscribeEvent
-    public static void onClientSetup(FMLClientSetupEvent event) {
-        // Some client setup code
-        WanderersOfTheRift.LOGGER.info("HELLO FROM CLIENT SETUP");
-        WanderersOfTheRift.LOGGER.info("MINECRAFT NAME >> {}", Minecraft.getInstance().getUser().getName());
-
-    }
-
-    @SubscribeEvent
-    public static void registerKeys(RegisterKeyMappingsEvent event) {
-        event.register(JIGSAW_NAME_TOGGLE_KEY);
-    }
-}
