+++ src/main/java/com/wanderersoftherift/wotr/init/client/ClientRegistryEvents.java
@@ -1,24 +1,32 @@
 package com.wanderersoftherift.wotr.init.client;
 
+import com.mojang.blaze3d.platform.InputConstants;
 import com.wanderersoftherift.wotr.WanderersOfTheRift;
 import com.wanderersoftherift.wotr.client.ModShaders;
 import com.wanderersoftherift.wotr.client.map.Direction;
 import com.wanderersoftherift.wotr.client.map.MapCell;
 import com.wanderersoftherift.wotr.client.map.MapData;
 import com.wanderersoftherift.wotr.client.map.MapRoom;
+import com.wanderersoftherift.wotr.client.render.blockentity.JigsawBlockEntityRenderer;
 import com.wanderersoftherift.wotr.client.render.entity.RiftEntranceRenderer;
+import com.wanderersoftherift.wotr.client.render.item.properties.select.SelectRuneGemShape;
+import com.wanderersoftherift.wotr.client.tooltip.GearSocketTooltipRenderer;
+import com.wanderersoftherift.wotr.client.tooltip.ImageTooltipRenderer;
+import com.wanderersoftherift.wotr.gui.layer.objective.ObjectiveLayer;
 import com.wanderersoftherift.wotr.gui.screen.KeyForgeScreen;
 import com.wanderersoftherift.wotr.gui.screen.RuneAnvilScreen;
 import com.wanderersoftherift.wotr.init.ModEntityTypes;
 import com.wanderersoftherift.wotr.init.ModMenuTypes;
+import net.minecraft.client.KeyMapping;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.neoforged.api.distmarker.Dist;
 import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.fml.common.EventBusSubscriber;
 import net.neoforged.fml.event.lifecycle.FMLClientSetupEvent;
-import net.neoforged.neoforge.client.event.EntityRenderersEvent;
-import net.neoforged.neoforge.client.event.RegisterMenuScreensEvent;
-import net.neoforged.neoforge.client.event.RegisterShadersEvent;
+import net.neoforged.neoforge.client.event.*;
 import org.joml.Vector3f;
+import org.lwjgl.glfw.GLFW;
 
 import java.util.ArrayList;
 import java.util.EnumSet;
@@ -67,11 +75,42 @@ public final class ClientRegistryEvents {
     @SubscribeEvent
     public static void registerEntityRenderers(EntityRenderersEvent.RegisterRenderers event) {
         event.registerEntityRenderer(ModEntityTypes.RIFT_ENTRANCE.get(), RiftEntranceRenderer::new);
+        event.registerBlockEntityRenderer(
+                BlockEntityType.JIGSAW,
+                JigsawBlockEntityRenderer::new
+        );
     }
 
     @SubscribeEvent
     public static void registerShaderPrograms(RegisterShadersEvent event) {
         event.registerShader(ModShaders.RIFT_PORTAL);
         event.registerShader(ModShaders.RIFT_MAPPER);
+    }
+    @SubscribeEvent
+    public static void registerSelectItemModelProperties(RegisterSelectItemModelPropertyEvent event) {
+        event.register(ResourceLocation.fromNamespaceAndPath(WanderersOfTheRift.MODID, "runegem_shape"), SelectRuneGemShape.TYPE);
+    }
+
+    @SubscribeEvent
+    public static void registerClientTooltipComponents(RegisterClientTooltipComponentFactoriesEvent event) {
+        event.register(ImageTooltipRenderer.ImageComponent.class, ImageTooltipRenderer::new);
+        event.register(GearSocketTooltipRenderer.GearSocketComponent.class, GearSocketTooltipRenderer::new);
+    }
+
+    @SubscribeEvent
+    public static void registerGuiLayers(RegisterGuiLayersEvent event) {
+        event.registerAboveAll(WanderersOfTheRift.id("objective"), new ObjectiveLayer());
+    }
+
+    public static final KeyMapping JIGSAW_NAME_TOGGLE_KEY = new KeyMapping(
+            "key." + WanderersOfTheRift.id("jigsaw_name_toggle"),
+            InputConstants.Type.KEYSYM,
+            GLFW.GLFW_KEY_UNKNOWN,
+            "key.categories.misc"
+    );
+
+    @SubscribeEvent
+    public static void registerKeys(RegisterKeyMappingsEvent event) {
+        event.register(JIGSAW_NAME_TOGGLE_KEY);
     }
 }
