+++ src/main/java/com/wanderersoftherift/wotr/network/ModPacketRegistrationEvents.java
@@ -1,7 +1,6 @@
-package com.wanderersoftherift.wotr.events;
+package com.wanderersoftherift.wotr.network;
 
 import com.wanderersoftherift.wotr.WanderersOfTheRift;
-import com.wanderersoftherift.wotr.network.S2CRiftObjectiveStatusPacket;
 import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.fml.common.EventBusSubscriber;
 import net.neoforged.neoforge.network.event.RegisterPayloadHandlersEvent;
@@ -16,6 +15,11 @@ public class ModPacketRegistrationEvents {
                 S2CRiftObjectiveStatusPacket.TYPE,
                 S2CRiftObjectiveStatusPacket.STREAM_CODEC,
                 new S2CRiftObjectiveStatusPacket.S2CRiftObjectiveStatusPacketHandler()
+        );
+        registrar.playToServer(
+                C2SRuneAnvilApplyPacket.TYPE,
+                C2SRuneAnvilApplyPacket.STREAM_CODEC,
+                new C2SRuneAnvilApplyPacket.C2SRuneAnvilApplyPacketHandler()
         );
     }
 }
